% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angell.R
\docType{data}
\name{Angell}
\alias{Angell}
\title{Angell}
\format{
This data frame contains the following columns:
\describe{
 \item{moral}{Moral Integration: Composite of crime rate and welfare expenditures.}
 \item{hetero}{Ethnic Heterogenity: From percentages of nonwhite andforeign-born white residents.}
 \item{mobility}{Geographic Mobility: From percentages of residents moving into and out of the city.}
 \item{region}{A factor with levels:
   \code{E} Northeast;
   \code{MW} Midwest;
   \code{S} Southeast;
   \code{W} West.
   }
 }
}
\source{
This data set is from package(carData)
}
\usage{
Angell
}
\description{
The data frame has 43 rows and 4 columns. The observations are 43 U. S. cities around 1950.
}
\keyword{datasets}
