% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatal}
\alias{fatal}
\title{Fatalism,}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Fatalism}{Fatalism score}
  \item{Age}{object's age }
  \item{R_E}{Race, Ethinity, non-Hi-American =1,  Mexican American = 0}
  \item{Sex}{Female=1, male=0}
  \item{Depression}{depression score}

  ...
}}
\usage{
fatal
}
\description{
A partial dataset from stroke study of 620 objects; measurement are collected on
their fatalism status, sex, age, race, and depression.
}
\keyword{datasets}
