% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLHtest.R
\name{GLHtest}
\alias{GLHtest}
\title{GLHtest}
\usage{
GLHtest(model, T, c = rep(0, nrow(T)))
}
\arguments{
\item{model}{input fitted linear model}

\item{T}{contrast matrix, each row specifies a linear combination of the coefficients}

\item{c}{vector, specifying the null hypothis values for each linear combination,
Default value is zero vector.}
}
\value{
F-statistics of the GLH test and corresponding p-value
}
\description{
General Linear Hypothesis Test
}
\examples{
n = 1e3
p = 20
Y = runif(n,-5,5)
X = matrix(rnorm(n*p,0,1),n,p)
T = matrix(sample(-5:5, p+1, replace = TRUE),nrow=1)
model = linearFit(X,Y)
GLHtest(model,T)

}
