% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent for ols.}
\usage{
gradient_descent(formula, dataframe, alpha = 0.01, num_iters = 10000,
  epsilon = 1e-08)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{dataframe}{a data frame containing the variables in the model.}

\item{alpha}{a float value representing learning rate. Default value is 0.1.
(default: contrasts.treatment)}

\item{num_iters}{the maximal number of iterations. Default value is 10000.}

\item{epsilon}{The measure of convergion. Default value is 1e-08.}
}
\value{
\code{gradient_descent} returns a vector of the coefficient of the fitted linear model
  derived through the method of ordinary least square, implemented by gradient descent method.
}
\description{
\code{gradient_descent} returns the coefficients of the linear model fiited.
}
\examples{
gradient_descent(Sepal.Length ~ ., iris, alpha = 0.01, num_iters = 150000, epsilon = 1e-08)
}
\keyword{descent}
\keyword{gradient}
