% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model.}
\usage{
linear_model(formula, dataframe, contrasts = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{dataframe}{a data frame containing the variables in the model.}

\item{contrasts}{an optional list that specifies a way to code a categorical variable.
(default: contrasts.treatment)}
}
\value{
\code{linear_model} returns an object of class 'list'. The list consists of 1
  component:
  \item{\code{coefficients}}{a named vector of coefficients}
}
\description{
\code{linear_model} returns the coefficients of the linear model fiited.
}
\examples{
lm(Sepal.Length  ~ ., iris)
lm(Sepal.Length  ~ ., iris, contrasts = list(Species = "contr.sum"))
}
\keyword{linear}
\keyword{model}
