% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearFit.R
\name{linearFit}
\alias{linearFit}
\title{linearFit}
\usage{
linearFit(X, Y)
}
\arguments{
\item{X}{input indicator variable matrix/design matrix}

\item{Y}{input response variable vector}
}
\value{
the coefficient of each indicator variable and corresponding
}
\description{
Gets the fitted linear regression model results
}
\examples{
n = 1e3
p = 20
Y = runif(n,-5,5)
X = matrix(rnorm(n*p,0,1),n,p)

}
