% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel-doc.R
\docType{package}
\name{linearmodel}
\alias{linearmodel}
\alias{linearmodel-package}
\title{linearmodel: A package for homework 1 of BIS 557 created by Xiaojing Zhu}
\description{
The linearmodel package provides 2 functions:
linear_model and gradient_descent
}
\section{linearmodel functions}{

The linear_model function implements the QR decomposition to obtain least
square estimates.

On the other hand, the gradient_descent function uses the gradient descent algorithm to obtain
least square estimates.
}

