% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{Experimental data set for fitting linear model}
\format{A data frame with 2 rows and 3 variables:
\describe{
  \item{y}{The numeric response variable}
  \item{x1}{The first numeric independent varaible}
  \iten{x2}{The second numeric independent varaible}
}}
\source{
\url{https://raw.githubusercontent.com/BIS557/homework-1-starter-code/master/lm_patho.csv}
}
\usage{
lm_patho
}
\description{
A dataset containing a response variable y and two independent varaibles x1, x2.
}
\keyword{datasets}
