% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmPredict.R
\name{lmPredict}
\alias{lmPredict}
\title{lmPredict}
\usage{
lmPredict(model, newX)
}
\arguments{
\item{model}{input fitted linear model}

\item{newX}{input indicator variable matrix,
one-dimension vector works as well}
}
\value{
the estimated Y-value
}
\description{
Use the fitted linear model to predict value for a new observation.
}
\examples{
n = 1e3
p = 20
Y = runif(n,-5,5)
X = matrix(rnorm(n*p,0,1),n,p)
newX = rnorm(p)
model = linearFit(X,Y)
lmPredict(model,newX)

}
