% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{lr}
\alias{lr}
\title{lr}
\usage{
lr(formula, data, intercept = T, interact = F, category = 3)
}
\arguments{
\item{formula}{the formula of your model}

\item{data}{the name of your data}

\item{intercept}{logical; if `TRUE`, intercept is included in your model}

\item{interact}{logical; if `TRUE`, interactions between variables are included in your model}

\item{category}{is the variable categorical in your model? Possible values are `1`, `2`, and `3`.
`1` means using the reference cell coding. `2` means using the cell means coding.
`3` is the default, meaning that the variable is continuous.}
}
\value{
`lr` returns a list of results regarding the linear model

y.fitted: a vector of fitted values of the dependent variables

residual: a matrix includes conventional residuals and internally studentized residuals

cov.matrix: a matrix includes the estimates, standard errors, t-test for each coefficient

df: the value of degrees of freedom in the model

r.squared: the value of r-squared in the model

sigma: the value of residual standard error in the model

f.result: a vector of F-test regarding the whole model
}
\description{
Generate linear regression model and perform basic analysis
}
\examples{
data(Angell)
# By default, the intercept is included, while the interactions are not.
lr(moral~hetero+mobility, data = Angell)

# When intercept = FALSE, the model would not include the intercept.
lr(moral~hetero+mobility, data = Angell, intercept = FALSE)

# When interaction = TRUE, the model would include the interactions between covariates.
lr(moral~hetero+mobility, data = Angell, interact = TRUE)

# Regarding the categorical variables, if category = 1, a reference cell coding is performed.
lr(moral~region, data = Angell, category = 1)

# If category = 2, then a cell means coding is performed.
lr(moral~region, data = Angell, category = 2)

}
