% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHANE.R
\docType{data}
\name{NHANE}
\alias{NHANE}
\title{NHANE}
\format{
This data frame contains the following columns:
\describe{
 \item{Age}{Age in years at screening of study participant.}
 \item{Race1}{Reported race of study participant: Mexican, Hispanic, White, Black, or Other.}
 \item{Weight}{Weight in kg.}
 \item{Height}{Standing height in cm. Reported for participants aged 2 years or older.}
 }
}
\source{
This raw data set is from package(NHANES)
}
\usage{
NHANE
}
\description{
The data frame has 9634 rows and 4 columns. This is the subset of a survey data collected by the US National Center for Health Statistics (NCHS) which has conducted a series of health and nutrition surveys since the early 1960’s.
}
\keyword{datasets}
