% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatal}
\alias{fatal}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Fatalism}{price, in US dollars}
  \item{Age}{weight of the diamond, in carats}
  \item{R_E}{Race, Ethinity}
  \item{Sex}{Female=1, male=0}
  \item{Depression}{depression score}

  ...
}}
\usage{
fatal
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
