% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Implement gradient descent for ordinary least squares.}
\usage{
gradient_descent(formula, data, contrasts = NULL, lambda = 1e-04,
  iter = 10000)
}
\arguments{
\item{formula}{An object of class formula to describe how the model is fitted.}

\item{data}{A data frame.}

\item{contrasts}{An optional list of contrasts for factor variables.}

\item{lambda}{A number for the learning rate, default = 0.0001.}

\item{iter}{A number for the number of iterations, default = 10000}
}
\value{
A list of vectors for coefficients, residuals and fitted.values.
}
\description{
Implement gradient descent for ordinary least squares.
}
\examples{
data(iris)
gradient_descent(Sepal.Length ~ ., iris, lambda = 0.001, iter = 3000000)
}
