% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Fit a linear model.}
\usage{
linear_model(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{An object of class formula to describe how the model is fitted.}

\item{data}{A data frame.}

\item{contrasts}{An optional list of contrasts for factor variables.}
}
\value{
A list of vectors for coefficients, residuals and fitted.values.
}
\description{
Fit a linear model.
}
\examples{
data(iris)
linear_model(Sepal.Length ~ ., iris)
}
