% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearFit.R
\name{linearFit}
\alias{linearFit}
\title{linearFit}
\usage{
linearFit(X, Y)
}
\arguments{
\item{X}{input indicator variable matrix, not including intercept 1}

\item{Y}{input response variable vector}
}
\value{
a list
$coefficient: the coefficient of each indicator variable and corresponding;
$test: single variable t-statistics and corresponding p-value
$CI: 95% confidence interval of coefficient estimate.
$MSE, $df.res
}
\description{
Gets the fitted linear regression model results
}
\examples{
n = 1e3
p = 20
Y = runif(n,-5,5)
X = matrix(rnorm(n*p,0,1),n,p)

}
