% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_patho-doc.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{lm_patho data within the linearmodel package}
\format{A data frame with 2 observations and 3 variables
\describe{
  \item{y}{The response variable}
  \item{x1}{First predictor}
  \item{x2}{Second predictor}
}}
\usage{
lm_patho
}
\description{
A dataset containing extreme cases for fitting a linear regression model
}
\keyword{datasets}
